Add files via upload
Cleanup, reworked some code that was unnecessarily repetitive, mainly adding a for loop on a list of all the tkinter widgets to disable instead of disabling them all individually. Also redid the keyboard bind listener to only use one function with an argument passsed of which button you are updating, instead of using 3 separate but identical functions.